-removed Icon* examples -moved SvgIcon to a separate directory -fixes
