--- packages/fluentui/docs/src/examples/components/Icon/Performance/IconMinimal.perf.tsx
@@ -1,9 +1,0 @@
-import { Icon } from '@fluentui/react-northstar';
-import * as React from 'react';
-
-const IconMinimalPerf = () => <Icon name="user" />;
-
-IconMinimalPerf.iterations = 5000;
-IconMinimalPerf.filename = 'IconMinimal.perf.tsx';
-
-export default IconMinimalPerf;
