--- packages/fluentui/docs/src/examples/components/Icon/Performance/index.tsx
@@ -1,16 +1,0 @@
-import * as React from 'react';
-
-import ComponentPerfExample from '../../../../components/ComponentDoc/ComponentPerfExample';
-import ExampleSection from '../../../../components/ComponentDoc/ExampleSection';
-
-const Performance = () => (
-  <ExampleSection title="Performance">
-    <ComponentPerfExample
-      title="Default"
-      description="A default test."
-      examplePath="components/Icon/Performance/IconMinimal.perf"
-    />
-  </ExampleSection>
-);
-
-export default Performance;
