--- packages/fluentui/docs/src/examples/components/Icon/Rtl/IconExample.rtl.tsx
@@ -1,13 +1,0 @@
-import * as React from 'react';
-import { Flex, Icon } from '@fluentui/react-northstar';
-
-const IconExampleRtl = () => (
-  <Flex gap="gap.smaller">
-    <Icon name="call-video" />
-    <Icon name="icon-circle" />
-    <Icon name="redo" />
-    <Icon name="send" />
-  </Flex>
-);
-
-export default IconExampleRtl;
