--- packages/fluentui/docs/src/examples/components/Icon/Rtl/index.tsx
@@ -1,13 +1,0 @@
-import * as React from 'react';
-
-import ComponentExample from '../../../../components/ComponentDoc/ComponentExample';
-import NonPublicSection from '../../../../components/ComponentDoc/NonPublicSection';
-
-const Rtl = () => (
-  <NonPublicSection title="Rtl">
-    <ComponentExample examplePath="components/Icon/Rtl/IconExample.rtl" />
-    <ComponentExample examplePath="components/Icon/Rtl/IconExampleRotate.rtl" />
-  </NonPublicSection>
-);
-
-export default Rtl;
