--- packages/fluentui/docs/src/examples/components/Icon/States/IconExampleDisabled.shorthand.tsx
@@ -1,10 +1,0 @@
-import * as React from 'react';
-import { Icon } from '@fluentui/react-northstar';
-
-const IconExampleDisabled = () => (
-  <div>
-    <Icon disabled name="call-video" />
-  </div>
-);
-
-export default IconExampleDisabled;
