--- packages/fluentui/docs/src/examples/components/Icon/States/index.tsx
@@ -1,15 +1,0 @@
-import * as React from 'react';
-import ComponentExample from '../../../../components/ComponentDoc/ComponentExample';
-import ExampleSection from '../../../../components/ComponentDoc/ExampleSection';
-
-const States = () => (
-  <ExampleSection title="States">
-    <ComponentExample
-      title="Disabled"
-      description="An icon can show it is currently unable to be interacted with."
-      examplePath="components/Icon/States/IconExampleDisabled"
-    />
-  </ExampleSection>
-);
-
-export default States;
