--- packages/fluentui/docs/src/examples/components/Icon/Types/IconExample.shorthand.tsx
@@ -1,13 +1,0 @@
-import * as React from 'react';
-import { Flex, Icon } from '@fluentui/react-northstar';
-
-const IconExample = () => (
-  <Flex gap="gap.smaller">
-    <Icon name="call-video" />
-    <Icon name="skype-logo" />
-    <Icon name="emoji" />
-    <Icon name="circle" />
-  </Flex>
-);
-
-export default IconExample;
