--- packages/fluentui/docs/src/examples/components/Icon/Types/index.tsx
@@ -1,11 +1,0 @@
-import * as React from 'react';
-import ComponentExample from '../../../../components/ComponentDoc/ComponentExample';
-import ExampleSection from '../../../../components/ComponentDoc/ExampleSection';
-
-const Types = () => (
-  <ExampleSection title="Types">
-    <ComponentExample title="Default" description="A default Icon." examplePath="components/Icon/Types/IconExample" />
-  </ExampleSection>
-);
-
-export default Types;
