--- packages/fluentui/docs/src/examples/components/Icon/Usage/index.tsx
@@ -1,15 +1,0 @@
-import * as React from 'react';
-import ComponentExample from '../../../../components/ComponentDoc/ComponentExample';
-import ExampleSection from '../../../../components/ComponentDoc/ExampleSection';
-
-const Usage = () => (
-  <ExampleSection title="Usage">
-    <ComponentExample
-      title="Theme Icons"
-      description="You can access all icon names available in the current theme."
-      examplePath="components/Icon/Usage/IconSetExample"
-    />
-  </ExampleSection>
-);
-
-export default Usage;
