--- packages/fluentui/docs/src/examples/components/Icon/Variations/IconExampleBordered.shorthand.tsx
@@ -1,17 +1,0 @@
-import * as React from 'react';
-import { Flex, Icon } from '@fluentui/react-northstar';
-
-const IconExampleBordered = () => (
-  <Flex gap="gap.smaller">
-    <Icon name="skype-logo" bordered />
-    <Icon name="emoji" bordered />
-    <Icon name="play" bordered />
-    <Icon name="call" bordered />
-    <Icon name="call-video" bordered />
-    <Icon name="calendar outline" bordered />
-    <Icon name="translation" bordered />
-    <Icon name="star outline" bordered />
-  </Flex>
-);
-
-export default IconExampleBordered;
