--- packages/fluentui/docs/src/examples/components/Icon/Variations/IconExampleCircular.shorthand.steps.ts
@@ -1,3 +1,0 @@
-const config: ScreenerTestsConfig = { themes: ['teams', 'teamsDark', 'teamsHighContrast'] };
-
-export default config;
