--- packages/fluentui/docs/src/examples/components/Icon/index.tsx
@@ -1,21 +1,0 @@
-import * as React from 'react';
-
-import Performance from './Performance';
-import Rtl from './Rtl';
-import Types from './Types';
-import Variations from './Variations';
-import States from './States';
-import Usage from './Usage';
-
-const IconExamples = () => (
-  <div>
-    <Types />
-    <States />
-    <Variations />
-    <Rtl />
-    <Performance />
-    <Usage />
-  </div>
-);
-
-export default IconExamples;
