+++ packages/fluentui/react-northstar/src/components/SvgIcon/SvgIcon.tsx
